+++ trunk-labwiz/LabWizFirmware/include/labwiz/drv_mcp3424.h
