--- trunk-labwiz/LabWizFirmware/include/thermocouples_private.h
@@ -1,211 +1,0 @@
-/*****************************************************************************
- **
- ** Thermocouple interface
- ** September 2016
- **
- ** This module provides functions related to thermocouples
- **
- ****************************************************************************/
-#if 0
-#ifndef __THERMOCOUPLES_PRIVATE_H__
-#define __THERMOCOUPLES_PRIVATE_H__
-
-// Macro for accessing thermocouple data
-// Usage: temperatureC = lookupThermocouleData(microvolts)
-//#define lookupThermocouleData(index)    (tempTypK[index]))
-//#define thermocoupleMicrovoltLookup(I)  lookupThermocouleData(I)
-
-// Note: if switching to eeprom, could look like this instead:
-//  #define lookupThermocouleData(index) (eeprom_read_dword(index));
-
-// Lookup Table for the TypK:
-// From -270 to 1350°C in steps of 10°C, the list is in µV according to that Temp.
-// If changing the lower limit remember to change it in GetTypKTemp() in functions.cpp.
-#define TEMP_TYPE_K_LENGTH 165 // Use the i values indicated below. TEMP_TYPE_K_LENGTH should be i+1
-
-
-// Note: if the table changes, update these by hand
-#define TEMP_TYPE_K_MIN_CONVERSION      0
-#define TEMP_TYPE_K_MAX_CONVERSION      61277
-
-// Add this to get all positive
-#define TK_OFFSET       6458
-// This is the lowest temp in the mv->table (deg C) (i=0)
-#define TK_MIN_TEMP     (-270)
-
-// Each entry is 10C apart
-// Usage: temperatureC = tempTypK[microvolts]
-int tempTypK[TEMP_TYPE_K_LENGTH] =
-{
-// Adjusted lookup voltage, normal lookup voltage, corresponding temperature
- 0,//,-6458,  //-270 i=0
-47,//,-6411, // -260
-54,//,-6404, // -250
-114,//,-6344, // -240
-196,//,-6262, // -230
-300,//,-6158, // -220
-423,//,-6035, // -210
-567,//,-5891, // -200
-728,//,-5730, // -190
-908,//,-5550, // -180
-1104,//,-5354, // -170 i=10
-1317,//,-5141, // -160
-1545,//,-4913, // -150
-1789,//,-4669, // -140
-2047,//,-4411, // -130
-2320,//,-4138, // -120
-2606,//,-3852, // -110
-2904,//,-3554, // -100
-3215,//,-3243, // -90
-3538,//,-2920, // -80
-3871,//,-2587, // -70 i=20
-4215,//,-2243, // -60
-4569,//,-1889, // -50
-4931,//,-1527, // -40
-5302,//,-1156,30
-5680,//,-778,20
-6066,//,-392,10
-6458,//,0,      // 0 // i=27
-6855,//,397,   // 10
-7256,//,798,   // 20
-7661,//,1203,  // 30 i=30
-8070,//,1612,  // 40
-8481,//,2023,  // 50
-8894,//,2436,  // 60
-9309,//,2851,  // 70
-9725,//,3267,  // 80
-10140,//,3682,  // 90
-10554,//,4096,  // 100
-10967,//,4509,  // 110
-11378,//,4920,  // 120
-11786,//,5328,  // 130 i=40
-12193,//,5735,  // 140
-12596,//,6138,  // 150
-12998,//,6540,  // 160
-13399,//,6941,  // 170
-13798,//,7340,  // 180
-14197,//,7739,  // 190
-14596,//,8138,  // 200
-14997,//,8539,  // 210
-15398,//,8940,  // 220
-15801,//,9343,  // 230 // i=50
-16205,//,9747,  // 240
-16611,//,10153, // 250
-17019,//,10561, // 260
-17429,//,10971, // 270
-17840,//,11382, // 280
-18253,//,11795, // 290
-18667,//,12209, // 300
-19082,//,12624, // 310
-19498,//,13040, // 320
-19915,//,13457, // 330 i=60
-20332,//,13874, // 340
-20751,//,14293, // 350
-21171,//,14713, // 360
-21691,//,15233, // 370
-22012,//,15554, // 380
-22433,//,15975, // 390
-22855,//,16397, // 400
-23278,//,16820, // 410
-23701,//,17243, // 420
-24125,//,17667, // 430 i=70
-24549,//,18091, // 440
-24974,//,18516, // 450
-25399,//,18941, // 460
-25824,//,19366, // 470
-26250,//,19792, // 480 // i = 75
-26586,//,20128, // 490
-27102,//,20644, // 500
-27529,//,21071, // 510
-27955,//,21497, // 520
-28382,//,21924, // 530 i=80
-28808,//,22350, // 540
-29234,//,22776, // 550
-29661,//,23203, // 560
-30087,//,23629, // 570
-30513,//,24055, // 580
-30938,//,24480, // 590
-31363,//,24905, // 600
-31788,//,25330, // 610
-32213,//,25755, // 620
-32637,//,26179, // 630 i=90
-33060,//,26602, // 640
-33483,//,27025, // 650
-33905,//,27447, // 660
-34327,//,27869, // 670
-34747,//,28289, // 680
-35168,//,28710, // 690
-35587,//,29129, // 700
-36006,//,29548, // 710
-36423,//,29965, // 720
-36840,//,30382, // 730 i=100
-37256,//,30798, // 740
-37671,//,31213, // 750
-38086,//,31628, // 760
-38499,//,32041, // 770
-38911,//,32453, // 780
-39323,//,32865, // 790
-39733,//,33275, // 800
-40143,//,33685, // 810
-40551,//,34093, // 820
-40959,//,34501, // 830 i=110
-41366,//,34908, // 840
-41771,//,35313, // 850
-42176,//,35718, // 860
-42579,//,36121, // 870
-42982,//,36524, // 880
-43383,//,36925, // 890
-43784,//,37326, // 900
-44183,//,37725, // 910
-44582,//,38124, // 920
-44980,//,38522, // 930 i=120
-45376,//,38918, // 940
-45772,//,39314, // 950
-46166,//,39708, // 960
-46559,//,40101, // 970
-46952,//,40494, // 980
-47343,//,40885, // 990
-47734,//,41276, // 1000
-48123,//,41665, // 1010
-48511,//,42053, // 1020
-48898,//,42440, // 1030 i=130
-49284,//,42826, // 1040
-49669,//,43211, // 1050
-50053,//,43595, // 1060
-50436,//,43978, // 1070
-50817,//,44359, // 1080
-51198,//,44740, // 1090
-51577,//,45119, // 1100
-51955,//,45497, // 1110
-52331,//,45873, // 1120
-52707,//,46249, // 1130 i=140
-53081,//,46623, // 1140
-53453,//,46995, // 1150
-53825,//,47367, // 1160
-54195,//,47737, // 1170
-54563,//,48105, // 1180
-54931,//,48473, // 1190
-55296,//,48838, // 1200
-55660,//,49202, // 1210
-56023,//,49565, // 1220
-56384,//,49926, // 1230 i=150
-56744,//,50286, // 1240
-57102,//,50644, // 1250
-57458,//,51000, // 1260
-57813,//,51355, // 1270
-58166,//,51708, // 1280
-58518,//,52060, // 1290
-58868,//,52410, // 1300
-59217,//,52759, // 1310
-59564,//,53106, // 1320
-59909,//,53451, // 1330 i=160
-60253,//,53795, // 1340
-60596,//,54138, // 1350
-60937,//,54479, // 1360
-61277,//,54819, // 1370 i=164
-};
-
-#endif
-#endif
-
-//eof
