+++ trunk-labwiz/LabWizFirmware/src/labwiz/drv_mcp9800.c
@@ -9,10 +9,9 @@
 
 #include "defs.h"
 #include "labwiz/drv_i2c.h"
+#include "labwiz/drv_mcp9800.h"
+
 #include "stm32f1xx_hal.h"
-
-#include "drv_mcp9800.h"
-
 #include "string.h"
 
 #ifdef ENABLE_MCP9800
